Add /people.
This commit adds the "/people" route. It represents a view of all registered people, their roles, and their keys in a single view. Users can see a list of members, choose one to load that person's details, and edit that person's personal info and access permissions.

This commit does NOT interface with the backend API.

> Adds Bootstrap, Lodash as dependencies
> Adds SearchBar, OptionsList, DetailsBoxForm, and DetailsBoxCheckboxes components
> Adds People view
> Adds /people route