Add /roles.
This commit adds /roles. The UI between /people and /roles is nearly identical up to the choice the details shown about a role. Whereas a Member has info, roles, and keys; a Role has info, doors, and people.

The mock API logic for roles is highly redundant with that for members. Deduplication coming soon...